+++ CotEditor/Sources/EditorViewController.swift
@@ -68,7 +68,7 @@ final class EditorViewController: NSSplitViewController {
         super.viewDidLoad()
         
         // workaround for OS X Yosemite (on macOS 10.12 SDK)
-        if NSAppKitVersionNumber < Double(NSAppKitVersionNumber10_11) {
+        if NSAppKitVersion.current < .macOS10_11 {
             self.splitView.delegate = self
         }