+++ CotEditor/Sources/FindPanelLayoutManager.swift
@@ -9,7 +9,7 @@
  
  ------------------------------------------------------------------------------
  
- © 2015-2016 1024jp
+ © 2015-2017 1024jp
  
  Licensed under the Apache License, Version 2.0 (the "License");
  you may not use this file except in compliance with the License.
@@ -86,7 +86,7 @@ final class FindPanelLayoutManager: NSLayoutManager {
             for glyphIndex in glyphsToShow.location..<glyphsToShow.upperBound {
                 let charIndex = self.characterIndexForGlyph(at: glyphIndex)
                 
-                let utfChar = string.utf16[String.UTF16Index(charIndex)]
+                let utfChar = string.utf16[String.UTF16Index(encodedOffset: charIndex)]
                 let character = String(utf16CodeUnits: [utfChar], count: 1)
                 
                 let glyphString: NSAttributedString