+++ CotEditor/Sources/WindowContentViewController.swift
@@ -45,7 +45,7 @@ final class WindowContentViewController: NSSplitViewController, TabViewControlle
         super.viewDidLoad()
         
         // workaround for OS X Yosemite (on macOS 10.12 SDK)
-        if NSAppKitVersionNumber < Double(NSAppKitVersionNumber10_11) {
+        if NSAppKitVersion.current < .macOS10_11 {
             self.splitView.delegate = self
         }
         
@@ -231,7 +231,7 @@ final class WindowContentViewController: NSSplitViewController, TabViewControlle
     /// set visibility and tab of sidebar
     private func setSidebarShown(_ shown: Bool, index: SidebarViewController.TabIndex? = nil, animate: Bool = false) {
         
-        guard NSAppKitVersionNumber >= Double(NSAppKitVersionNumber10_11) else {
+        guard NSAppKitVersion.current >= .macOS10_11 else {
             self.isSidebarShown = shown
             return
         }